+++ packages/xandr_ios/ios/Classes/models/FlutterState.swift
@@ -108,4 +108,16 @@ public class FlutterState {
       completion: logResult
     )
   }
+
+  public func onNativeAdLoadedAPI(viewId: Int64, title: String, description: String,
+                                  imageUrl: String, clickUrl: String) {
+    flutterAPI?.onNativeAdLoaded(
+      viewId: viewId,
+      title: title,
+      description: description,
+      imageUrl: imageUrl,
+      clickUrl: clickUrl,
+      completion: logResult
+    )
+  }
 }