+++ packages/xandr_ios/pigeons/messages.dart
@@ -37,6 +37,7 @@ abstract class XandrHostApi {
   bool initXandrSdk({
     required int memberId,
     int? publisherId,
+    bool testMode = false,
   });
 
   @async
@@ -106,6 +107,7 @@ abstract class XandrFlutterApi {
     String title,
     String description,
     String imageUrl,
+    String clickUrl,
   ) {}
   void onNativeAdLoadedError(int viewId, String reason) {}
   void onAdClicked(int viewId, String url) {}