+++ packages/xandr_platform_interface/lib/src/interface.dart
@@ -39,8 +39,16 @@ abstract class XandrPlatform extends PlatformInterface {
   }
 
   /// Return the current platform name.
-  Future<bool> init(int memberId, {int? publisherId}) {
-    return _instance.init(memberId, publisherId: publisherId);
+  Future<bool> init(
+    int memberId, {
+    int? publisherId,
+    bool testMode = false,
+  }) {
+    return _instance.init(
+      memberId,
+      publisherId: publisherId,
+      testMode: testMode,
+    );
   }
 
   /// loads an ad.