+++ packages/xandr_android/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 0.2.7
+
+ - Fix null check in kotlin land
+
+ - **FIX**: null safe call on xandr native ad response.
+ - **FEAT**: add test mode config param on init.
+
+## 0.2.6
+
+ - **FEAT**: clickUrl implementation on nativeAdLoaded event.
+
 ## 0.2.5+2
 
  - **FIX**: fix clickThroughAction ios implementation + onAdClicked callback.