+++ packages/xandr_platform_interface/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.2.7
+
+ - **FEAT**: add test mode config param on init.
+
+## 0.2.6
+
+ - **FEAT**: clickUrl implementation on nativeAdLoaded event.
+
 ## 0.2.5+2
 
  - **FIX**: fix clickThroughAction ios implementation + onAdClicked callback.